Fixed log-in log-out problem using cookies.

+++ delete_p.php
@@ -17,6 +17,16 @@
     include("functions/alert.php");
 
     $uname= $_GET['uname'];
+    if (isset($_COOKIE['admin'])) {
+        $admin = $_COOKIE['admin'];
+    } else {
+        $admin = "";
+    }
+    if ($admin == NULL || $admin != $uname) {
+      setAlert("Please log in.");
+      echo "<div class='text-center'><a href='login.php?' class='btn btn-success btn-lg' role='button'>Go Log in!</a></div>";
+      die;
+    }
     $mid = $_GET['mid'];
 
     $sql="delete from post_m where mid = '{$mid}' and username = '{$uname}';";
